New look-into-the-future algorithm
Instead of returning a score from the most probable board (which was
probably incorrectly implemented anyway), these changes return the
actual board of the most probable state of the game starting with a
given legal move.

Starting at the bottom of the recursive get_final_board_state(...)
stack, the method simply returns the board after the move has been
submitted. At higher levels, the board with the highest rating according
to which side is making the move is returned. These can come from
different game depths if some lines have fewer legal moves than others
(thinner branches on the game tree). Finally, when the evaluation makes
it back to the choose_move(...) method, it picks the move that leads to
the best board in the future (as far as it can see).

+++ include/Genes/Look_Ahead_Gene.h
@@ -19,7 +19,7 @@ class Look_Ahead_Gene : public Gene
 
         std::string name() const override;
 
-        size_t positions_to_examine(const Board& board, const Clock& clock) const;
+        double positions_to_examine(const Board& board, const Clock& clock) const;
 
     private:
         void reset_properties() const override;
